requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pandas==2.1.3  # For efficient CSV processing
requests==2.31.0  # For webhook HTTP requests

# Development
django-extensions==3.2.3

